fix: prevent page from flickering on load
if shouldRender is false, it prevents the page from rendering. By default it was set to false and then after the delay it was set to true. When there was no transition, the delay would be set to 0. This caused occasional flickering on first load. It wasn't noticeable until persistent footers were introduced. This fixes that issue